Ahereum
Ahereum aims to provide a platform, based on an authorized blockchain, to save, immutably, the path of a drone or a flock of drones. The ultimate goal would be to integrate a simulation, based on the Digital Twin, to test the proper integration of the drone with the data saved in the blockchain.